Minor style fixes

+++ src/parquet/column/column-writer-test.cc
@@ -190,7 +190,7 @@ void TestPrimitiveWriter<FLBAType>::ReadColumnFully(Compression::type compressio
   this->data_buffer_.clear();
 
   values_read_ = 0;
-  while (values_read_ < this->values_out_.size()) {
+  while (values_read_ < static_cast<int64_t>(this->values_out_.size())) {
     int64_t values_read_recently = 0;
     reader_->ReadBatch(this->values_out_.size() - values_read_,
         definition_levels_out_.data() + values_read_,
